backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
python-multipart==0.0.17

